added changes for clone app logic

+++ builder/app-builder-upgrade/app-builder-upgrade.service.ts
@@ -565,8 +565,8 @@ export class AppBuilderUpgradeService {
     private async uninstallWidgets(plugins: any) {
         const appList = await this.getApplicationList();
         for (const pluginBinary of plugins) {
-            const widgetAppObj = appList.find(app => pluginBinary.oldContextPath ? app.contextPath === pluginBinary.oldContextPath : app.contextPath === pluginBinary.contextPath)
-            if (widgetAppObj) {
+            const widgetAppObj: IApplication = appList.find(app => pluginBinary.oldContextPath ? app.contextPath === pluginBinary.oldContextPath : app.contextPath === pluginBinary.contextPath)
+            if (widgetAppObj  && (widgetAppObj.manifest &&  !widgetAppObj.manifest?.isPackage)) {
                 try {
                     await this.appService.delete(widgetAppObj.id);
                 } catch (e) {
